+++ kutana/__init__.py
@@ -6,7 +6,7 @@ from .loaders import import_module, load_plugins, load_plugins_from_file
 from .utils import get_path, is_list_or_tuple, sort_callbacks
 from .kutana import Kutana
 from .logger import logger, set_logger_level
-from .manager.basic import BasicManager
+from .manager.manager import Manager
 from .manager.debug import DebugEnvironment, DebugManager
 from .manager.tg import TGEnvironment, TGManager, TGResponse
 from .manager.vk import VKEnvironment, VKManager, VKRequest, VKResponse